feat(backoffice): unify ticketing steps into master-detail checkout builder
Merge the listing, add-step modal and full-page editor into one
master-detail surface at /ticketing-steps. A left journey rail selects
a step via a ?step search param; the right pane shows the step detail,
an inline new-step panel, or an empty state.

- Extract the duplicated template grid into a shared TemplatePicker
- Relocate the editor into StepDetailPanel (save keeps the panel open)
- Replace AddStepDialog with an inline NewStepPanel
- Redirect the old $stepId route to the search-param selection
