+++ backoffice/src/routes/_layout/ticketing-steps/index.tsx
@@ -14,8 +14,9 @@ import { useEffect, useMemo, useState } from "react"
 
 import { type TicketingStepPublic, TicketingStepsService } from "@/client"
 import { WorkspaceAlert } from "@/components/Common/WorkspaceAlert"
-import { AddStepDialog } from "@/components/ticketing-step-builder/AddStepDialog"
+import { NewStepPanel } from "@/components/ticketing-step-builder/NewStepPanel"
 import { StepCanvas } from "@/components/ticketing-step-builder/StepCanvas"
+import { StepDetailPanel } from "@/components/ticketing-step-builder/StepDetailPanel"
 import { Button } from "@/components/ui/button"
 import { Skeleton } from "@/components/ui/skeleton"
 import { useWorkspace } from "@/contexts/WorkspaceContext"
@@ -23,8 +24,15 @@ import useAuth from "@/hooks/useAuth"
 import useCustomToast from "@/hooks/useCustomToast"
 import { createErrorHandler } from "@/utils"
 
+interface TicketingStepsSearch {
+  step?: string
+}
+
 export const Route = createFileRoute("/_layout/ticketing-steps/")({
   component: TicketingStepsPage,
+  validateSearch: (raw: Record<string, unknown>): TicketingStepsSearch => ({
+    ...(typeof raw.step === "string" && raw.step ? { step: raw.step } : {}),
+  }),
   head: () => ({
     meta: [{ title: "Ticketing Steps - EdgeOS" }],
   }),
@@ -59,9 +67,10 @@ function TicketingStepsPage() {
 function TicketingStepsContent({ popupId }: { popupId: string }) {
   const queryClient = useQueryClient()
   const navigate = useNavigate()
+  const { step } = Route.useSearch()
   const { showErrorToast } = useCustomToast()
 
-  const [addDialogOpen, setAddDialogOpen] = useState(false)
+  const [adding, setAdding] = useState(false)
   const [displayOrder, setDisplayOrder] = useState<string[]>([])
 
   const { data: stepsData, isLoading } = useQuery({
@@ -129,12 +138,22 @@ function TicketingStepsContent({ popupId }: { popupId: string }) {
     persistStepOrder(newOrder)
   }
 
-  const handleEdit = (step: TicketingStepPublic) => {
-    navigate({
-      to: "/ticketing-steps/$stepId",
-      params: { stepId: step.id },
-    })
+  const selectStep = (id: string) => {
+    setAdding(false)
+    navigate({ to: "/ticketing-steps", search: { step: id } })
   }
+
+  const clearSelection = () => {
+    setAdding(false)
+    navigate({ to: "/ticketing-steps", search: {} })
+  }
+
+  const startAdding = () => {
+    setAdding(true)
+    navigate({ to: "/ticketing-steps", search: {} })
+  }
+
+  const selectedStep = step ? steps.find((s) => s.id === step) : undefined
 
   if (isLoading) {
     return (
@@ -151,8 +170,9 @@ function TicketingStepsContent({ popupId }: { popupId: string }) {
   }
 
   return (
-    <div className="flex flex-col gap-6">
-      <div className="flex items-start justify-between gap-4">
+    <div className="flex flex-col gap-6 md:flex-row md:items-start">
+      {/* Journey rail */}
+      <div className="flex w-full flex-col gap-4 md:w-80 md:max-w-sm md:shrink-0">
         <div>
           <h1 className="text-2xl font-bold tracking-tight">Ticketing Steps</h1>
           <p className="text-muted-foreground">
@@ -160,17 +180,45 @@ function TicketingStepsContent({ popupId }: { popupId: string }) {
             rename
           </p>
         </div>
-        <Button onClick={() => setAddDialogOpen(true)}>+ Add Step</Button>
-      </div>
 
-      <div className="max-w-xl">
         <DndContext
           sensors={sensors}
           collisionDetection={closestCenter}
           onDragEnd={handleDragEnd}
         >
-          <StepCanvas steps={orderedSteps} onEdit={handleEdit} />
+          <StepCanvas
+            steps={orderedSteps}
+            onEdit={(s) => selectStep(s.id)}
+            selectedId={selectedStep?.id}
+          />
         </DndContext>
+
+        <Button variant="outline" onClick={startAdding}>
+          + Add step
+        </Button>
+      </div>
+
+      {/* Detail pane */}
+      <div className="min-w-0 flex-1">
+        {adding ? (
+          <NewStepPanel
+            popupId={popupId}
+            nextOrder={orderedSteps.length}
+            confirmStepId={steps.find((s) => s.step_type === "confirm")?.id}
+            onCreated={(id) => selectStep(id)}
+            onCancel={clearSelection}
+          />
+        ) : selectedStep ? (
+          <StepDetailPanel
+            key={selectedStep.id}
+            stepId={selectedStep.id}
+            onClose={clearSelection}
+          />
+        ) : (
+          <div className="rounded-lg border border-dashed p-10 text-center text-sm text-muted-foreground">
+            Select a step to configure it, or add a new one
+          </div>
+        )}
       </div>
 
       {updateStepMutation.isPending && (
@@ -179,14 +227,6 @@ function TicketingStepsContent({ popupId }: { popupId: string }) {
           Saving...
         </div>
       )}
-
-      <AddStepDialog
-        open={addDialogOpen}
-        onOpenChange={setAddDialogOpen}
-        popupId={popupId}
-        nextOrder={orderedSteps.length}
-        confirmStepId={steps.find((s) => s.step_type === "confirm")?.id}
-      />
     </div>
   )
 }
